reference commit on image model
